requirements.txt
Django==5.2.8
psycopg2-binary==2.9.9
python-docx==1.1.2
Pillow==10.4.0
gunicorn==21.2.0
whitenoise==6.6.0
dj-database-url==2.1.0
python-decouple==3.8

